app/requirements.txt
slack-bolt==1.18.1
slack_sdk==3.27.1
python-dotenv==1.0.0